refactor: MobileHeader 분리
